Update script.js
We improved the way the application switches between SFW and NSFW modes. Now, when you switch between these modes, the tags in the dropdown menu are updated automatically to match the selected mode. And, a new image is fetched and displayed based on the tags you choose and the mode you're in. These changes were made to ensure the application works smoothly and provides a better user experience.

When StevenK-293 updated the code, there was a minor issue, but fortunately, it was an easy fix. and Everything is now working. :)